[TASK] Refactor for usage outside TYPO3 instance
This commit adds the possibility of adding storages outside public with
having speaking file names in URL. To do so, the getPublicUrl() event is
used to manipulate the public URL of a non-public file storage.
With this, a real path is generated and not the eId for dumpFile from
TYPO3 core is used. This allows a fine granulated trigger to the found
storages, and checks access real with a 403 response if no logged-in
user is set.