[TASK] Register solr modify event listener via `Services.yaml`
Until now the PSR-14 `PreAddModifyDocuments` event listener
has been registered in the extension `Services.php` file to
make this conditional based on the availbility of the event
provided by the optional `apache-solr-for-typo3/solr` TYPO3
extension.

PSR-14 events are naturally plain `DTO's` and are not really
covered by the DI container, which means that the event may
be loaded first but removed from DI container because DTO's
are instantiated using `new` and are not DI aware - you know,
DTO`s are not services, right?

This change moves the PSR-14 event listener registration to
the `Services.yaml`, simple register for the optional event
`BeforeDocumentIsProcessedForIndexingEvent` and let the TYPO3
event dispatcher (ListenerProvider) do the safe check, which
is based on class-name checks and safe even when the class
does not exists. Simple said, let TYPO3 make the check and
reduce own code handling.