[BUGFIX] Check for baseUri array key before trying to access it
Because trying to access the baseUri without the array key
present will result in the frontend breaking completely down due
to "Trying to access a non-existent array key", the array key is
now checked before accessing it. This happens if the baseUri is
not set in a publicly non-available but online file storage. As this
sometimes can be intentional but the `EXT:typo3-secure-filemount`
depends on it to be set for handling secure files the
check is neccessary to ensure the extension functionality and
prevent breaking in environments where a file storage has met
the conditions but without having a baseUri set.